requirements.txt
hl7apy>=1.3.4
pydicom>=2.4.4
pynetdicom>=2.0.0
fastapi>=0.111.0
uvicorn>=0.30.1
httpx>=0.27.0
backoff>=2.2.1
python-dotenv>=1.0.1
pydantic>=2.6.0
pdfminer.six>=20231228
reportlab>=4.2.2
python-json-logger>=3.3.0
prometheus-fastapi-instrumentator>=7.1.0
prometheus-client>=0.22.1
PyJWT>=2.8.0
bcrypt>=4.1.2
python-jose>=3.3.0
passlib>=1.7.4

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Data Generation & Caching
faker>=24.0.0
redis>=5.0.0

# WebSocket & Real-time
python-socketio>=5.11.0
python-engineio>=4.9.0

# Scheduling & Background Tasks
apscheduler>=3.10.0

# Logging
structlog>=24.1.0

# NST Integration Dependencies
matplotlib>=3.7.0
numpy>=1.24.0
pyserial>=3.5
pytest>=8.0.0
respx>=0.21.1
pyusb>=1.2.1